Adding typecast to remove warnings from gcc mingw version 3.4.5 on Windows.

+++ applications/desktop_sniffer/sniff_parse/on_packet.cpp
@@ -2206,7 +2206,7 @@ on_packet::on_packet(std::string encoded_packet, std::string key)
 
     decoded_payload = bytes_to_hex_string(decoded_payload_bytes, decoded_payload_len);
 
-    if((raw_pid & 0x3F) < NUM_PIDS)
+    if((UInt16)(raw_pid & 0x3F) < NUM_PIDS)
     {
         valid_pid = true;
     }
